add common summary component; update changelog
