+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add Infrastructure as Code (IaC) support
 - Add Secrets Violation Card
+- Add IaC Violation Card
 - Add icons for file nodes in the tree view
 - Add a full path of file nodes in the tree view
