+++ src/main/kotlin/com/cycode/plugin/components/toolWindow/components/violationCardContentTab/secretViolationCardContentTab/components/summary/SecretSummary.kt
@@ -2,39 +2,18 @@ package com.cycode.plugin.components.toolWindow.components.violationCardContentT
 
 import com.cycode.plugin.CycodeBundle
 import com.cycode.plugin.cli.models.scanResult.secret.SecretDetection
-import com.intellij.ui.components.JBLabel
-import com.intellij.util.ui.JBFont
-import com.intellij.util.ui.JBUI
-import java.awt.GridBagConstraints
-import java.awt.GridBagLayout
+import com.cycode.plugin.components.toolWindow.components.violationCardContentTab.common.summary.CardSummary
 import javax.swing.JComponent
-import javax.swing.JPanel
 
-class SecretSummary {
+class SecretSummary : CardSummary() {
     fun getContent(detection: SecretDetection): JComponent {
-        val gbc = GridBagConstraints()
-        gbc.fill = GridBagConstraints.HORIZONTAL
-        gbc.anchor = GridBagConstraints.NORTHWEST
-        gbc.weightx = 1.0
-
-        val panel = JPanel(GridBagLayout())
-
-        gbc.gridy = 1
-        gbc.insets = JBUI.insetsBottom(5)
-        panel.add(JBLabel(CycodeBundle.message("secretViolationCardSummaryTitle")).apply {
-            font = font.deriveFont(18f).deriveFont(JBFont.BOLD)
-        }, gbc)
-
-        gbc.gridy++
-        gbc.insets = JBUI.emptyInsets()
-
+        val title = CycodeBundle.message("secretViolationCardSummaryTitle")
         val fixedMessage = detection.message.replace(
             "within '' repository", // BE bug
             ""
         )
         val message = detection.detectionDetails.description ?: fixedMessage
-        panel.add(JBLabel(message).apply { setAllowAutoWrapping(true); setCopyable(true) }, gbc)
 
-        return panel
+        return getContent(title, message)
     }
 }
